@@ -9,13 +9,13 @@ namespace SpeakerMeet.Core.Entities
         [Key]
         public Guid Id { get; set; }
 
-        public string Name { get; set; } = null!;
+        public string Name { get; set; } = default!;
 
-        public string Slug { get; set; } = null!;
+        public string Slug { get; set; } = default!;
 
-        public string Location { get; set; } = null!;
+        public string Location { get; set; } = default!;
 
-        public string Description { get; set; } = null!;
+        public string Description { get; set; } = default!;
 
         public bool IsActive { get; set; }
 
